mcp-server/requirements.txt
fastapi==0.110.1
uvicorn==0.27.1
openai==0.28.1
requests==2.31.0
python-dotenv==1.0.1